feat: make camera follow player, camera transition, camera shake

+++ Assets/_Scripts/Events/Events.cs
@@ -2,7 +2,7 @@ using UnityEngine.Events;
 
 public static class PlayerEvents
 {
-    public static UnityAction<float> HpChanged;
+    public static UnityAction<int> HpChanged;
     public static UnityAction Killed;
     public static UnityAction Spawned;
     // etc etc
